+++ src/activations/tanh.c
@@ -5,13 +5,13 @@ extern __Model__ * m;
 
 dARRAY * forward_pass_tanh(){
   dARRAY * tanh_out = (dARRAY*)malloc(sizeof(dARRAY));
-  tanh_out->matrix = (double*)calloc(m->current_layer->DENSE->cache->shape[0]*m->current_layer->DENSE->cache->shape[1],sizeof(double));
+  tanh_out->matrix = (float*)calloc(m->current_layer->DENSE->cache->shape[0]*m->current_layer->DENSE->cache->shape[1],sizeof(float));
   omp_set_num_threads(8);
   #pragma omp parallel for num_threads(8) shared(m,tanh_out) schedule(static)
   for(int i=0;i<m->current_layer->DENSE->cache->shape[0]*m->current_layer->DENSE->cache->shape[1];i++){
     //Computing the tanh function
-    double exp_res1 = exp(m->current_layer->DENSE->cache->matrix[i]);
-    double exp_res2 = exp(-1*m->current_layer->DENSE->cache->matrix[i]);
+    float exp_res1 = exp(m->current_layer->DENSE->cache->matrix[i]);
+    float exp_res2 = exp(-1*m->current_layer->DENSE->cache->matrix[i]);
     tanh_out->matrix[i] = (exp_res1 - exp_res2)/(exp_res1 + exp_res2);
   }
   tanh_out->shape[0] = m->current_layer->DENSE->cache->shape[0];
@@ -21,14 +21,14 @@ dARRAY * forward_pass_tanh(){
 
 dARRAY * backward_pass_tanh(){
   dARRAY * tanh_out = (dARRAY*)malloc(sizeof(dARRAY));
-  tanh_out->matrix = (double*)calloc(m->current_layer->DENSE->A->shape[0]*m->current_layer->DENSE->A->shape[1],sizeof(double));
+  tanh_out->matrix = (float*)calloc(m->current_layer->DENSE->A->shape[0]*m->current_layer->DENSE->A->shape[1],sizeof(float));
   omp_set_num_threads(8);
   #pragma omp parallel for num_threads(8) shared(m,tanh_out) schedule(static)
   for(int i=0;i<m->current_layer->DENSE->A->shape[0]*m->current_layer->DENSE->A->shape[1];i++){
     //Computing the tanh function
-    double exp_res1 = exp(m->current_layer->DENSE->A->matrix[i]);
-    double exp_res2 = exp(-1*m->current_layer->DENSE->A->matrix[i]);
-    double temp = (exp_res1 - exp_res2)/(exp_res1 + exp_res2);
+    float exp_res1 = exp(m->current_layer->DENSE->A->matrix[i]);
+    float exp_res2 = exp(-1*m->current_layer->DENSE->A->matrix[i]);
+    float temp = (exp_res1 - exp_res2)/(exp_res1 + exp_res2);
     //gradient of tanh is g'(z) = 1 - (tanh(z))^2
     tanh_out->matrix[i] = 1-pow(temp,2);
   }
